Rename imports of matrix package classes
