+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/layer/pathfinding/PathfindingLayer.java
@@ -8,10 +8,10 @@ import java.util.List;
 import org.firstinspires.ftc.teamcode.Units;
 import org.firstinspires.ftc.teamcode.layer.Layer;
 import org.firstinspires.ftc.teamcode.layer.LayerSetupInfo;
-import org.firstinspires.ftc.teamcode.localization.Mat2;
-import org.firstinspires.ftc.teamcode.localization.Mat3;
+import org.firstinspires.ftc.teamcode.matrix.Mat2;
+import org.firstinspires.ftc.teamcode.matrix.Mat3;
 import org.firstinspires.ftc.teamcode.localization.RobotLocalizer;
-import org.firstinspires.ftc.teamcode.localization.Vec2;
+import org.firstinspires.ftc.teamcode.matrix.Vec2;
 import org.firstinspires.ftc.teamcode.task.HolonomicDriveTask;
 import org.firstinspires.ftc.teamcode.task.MoveToFieldTask;
 import org.firstinspires.ftc.teamcode.task.Task;
